+++ src/Sylius/Behat/Resources/config/services/contexts/setup.xml
@@ -90,6 +90,7 @@
             <argument type="service" id="sylius.repository.customer" container="symfony" />
             <argument type="service" id="sylius.order_processing.order_recalculator" container="symfony" />
             <argument type="service" id="doctrine.orm.entity_manager" container="symfony" />
+            <argument type="service" id="sm.factory" container="symfony" />
             <tag name="sylius.behat.context" />
         </service>
