+++ src/Sylius/Bundle/CoreBundle/Doctrine/ORM/PaymentRepository.php
@@ -12,11 +12,28 @@
 namespace Sylius\Bundle\CoreBundle\Doctrine\ORM;
 
 use Pagerfanta\Pagerfanta;
-use Sylius\Bundle\PaymentBundle\Doctrine\ORM\PaymentRepository as BasePaymentRepository;
+use Sylius\Bundle\ResourceBundle\Doctrine\ORM\EntityRepository;
 use Sylius\Component\Core\Repository\PaymentRepositoryInterface;
 
-class PaymentRepository extends BasePaymentRepository implements PaymentRepositoryInterface
+class PaymentRepository extends EntityRepository implements PaymentRepositoryInterface
 {
+    /**
+     * {@inheritdoc}
+     */
+    public function findByOrderIdAndId($orderId, $id)
+    {
+        $queryBuilder = $this->createQueryBuilder('o');
+
+        return $queryBuilder
+            ->where('o.order = :orderId')
+            ->andWhere('o.id = :id')
+            ->setParameter('orderId', $orderId)
+            ->setParameter('id', $id)
+            ->getQuery()
+            ->getOneOrNullResult()
+        ;
+    }
+
     /**
      * @param array $criteria
      * @param array $sorting
