+++ src/Sylius/Component/Core/Repository/PaymentRepositoryInterface.php
@@ -12,12 +12,13 @@
 namespace Sylius\Component\Core\Repository;
 
 use Pagerfanta\Pagerfanta;
-use Sylius\Component\Payment\Repository\PaymentRepositoryInterface as BasePaymentRepositoryInterface;
+use Sylius\Component\Core\Model\PaymentInterface;
+use Sylius\Component\Resource\Repository\RepositoryInterface;
 
 /**
  * @author Łukasz Chruściel <lukasz.chrusciel@lakion.com>
  */
-interface PaymentRepositoryInterface extends BasePaymentRepositoryInterface
+interface PaymentRepositoryInterface extends RepositoryInterface
 {
     /**
      * @param array|null $criteria
@@ -26,4 +27,12 @@ interface PaymentRepositoryInterface extends BasePaymentRepositoryInterface
      * @return Pagerfanta
      */
     public function createFilterPaginator(array $criteria = null, array $sorting = null);
+
+    /**
+     * @param mixed $orderId
+     * @param mixed $id
+     *
+     * @return PaymentInterface|null
+     */
+    public function findByOrderIdAndId($orderId, $id);
 }
